chore: matcher move to lib folder
